webpack-configuration-part-three

+++ Wetube/src/controllers/videoController.js
@@ -8,27 +8,27 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params
-  const {
-    user: { _id },
-  } = req.session
   const video = await Video.findById(id).populate('owner')
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found.' })
   }
-  if (String(video.owner) !== String(_id)) {
-    return res.status(403).redirect('/')
-  }
-  return res.render('edit', {
-    pageTitle: `Edit:${video.title}`,
+  return res.render('watch', {
+    pageTitle: video.title,
     video,
   })
 }
 
 export const getEdit = async (req, res) => {
   const { id } = req.params
+  const {
+    user: { _id },
+  } = req.session
   const video = await Video.findById(id)
   if (!video) {
-    return res.render('404', { pageTitle: 'Video not found.' })
+    return res.status('404', { pageTitle: 'Video not found.' })
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/')
   }
   return res.render('edit', { pageTitle: `Edit: ${video.title}`, video })
 }
@@ -87,6 +87,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params
+  const {
+    user: { _id },
+  } = req.session
+  const video = await Video.findById(id)
+  if (!video) {
+    return res.status(404).render('404', { pageTitle: 'Video not found..' })
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/')
+  }
   await Video.findByIdAndDelete(id)
   return res.redirect('/')
 }
